New server.rs module
