+++ apps/render_cop/src/main.rs
@@ -4,10 +4,8 @@
 use anyhow::{anyhow, Context, Result};
 use hapi_rs::asset::AssetLibrary;
 use hapi_rs::node::{HoudiniNode, NodeFlags, NodeType};
-use hapi_rs::{
-    parameter::*,
-    session::{new_in_process_session, SessionOptions},
-};
+use hapi_rs::session::{SessionOptions, new_in_process_session};
+use hapi_rs::{parameter::*};
 use iced::keyboard::KeyCode::Space;
 use iced::widget::image;
 use iced::widget::pane_grid::Axis::Horizontal;
