+++ lib/examples/curve_marshall.rs
@@ -1,5 +1,7 @@
 /// curve_marshall.cpp
-use hapi_rs::{Result, attribute::*, geometry::*, session::*};
+use hapi_rs::{Result, attribute::*, geometry::*};
+use hapi_rs::server::ServerOptions;
+use hapi_rs::session::{SessionOptions, new_thrift_session};
 
 fn main() -> Result<()> {
     let session = new_thrift_session(SessionOptions::default(), ServerOptions::shared_memory())?;
