+++ lib/examples/materials.rs
@@ -2,7 +2,8 @@ use hapi_rs::Result;
 /// Example extracts material textures to files
 use hapi_rs::geometry::Materials;
 use hapi_rs::parameter::*;
-use hapi_rs::session::{ServerOptions, SessionOptions, new_thrift_session};
+use hapi_rs::server::ServerOptions;
+use hapi_rs::session::{SessionOptions, new_thrift_session};
 
 fn main() -> Result<()> {
     let session = new_thrift_session(SessionOptions::default(), ServerOptions::shared_memory())?;
