+++ lib/examples/packed_primitives.rs
@@ -4,7 +4,8 @@
 use hapi_rs::Result;
 use hapi_rs::geometry::{PackedPrimInstancingMode as IM, *};
 use hapi_rs::node::*;
-use hapi_rs::session::*;
+use hapi_rs::server::ServerOptions;
+use hapi_rs::session::{CookOptions, SessionOptions, new_thrift_session};
 
 fn main() -> Result<()> {
     let session = new_thrift_session(SessionOptions::default(), ServerOptions::shared_memory())?;
