+++ lib/src/parameter/mod.rs
@@ -4,7 +4,8 @@
 //! use pattern matching:
 //!
 //! ```
-//! use hapi_rs::session::{new_in_process_session, SessionOptions};
+//! use hapi_rs::session::new_in_process_session;
+//! use hapi_rs::session::SessionOptions;
 //! use hapi_rs::parameter::*;
 //! let session = new_in_process_session(Some(SessionOptions::default())).unwrap();
 //! let lib = session.load_asset_file("../otls/hapi_parms.hda").unwrap();
